requirements.txt
langchain-openai
langchain
langchain-community
faiss-cpu
tqdm
python-dotenv
PyPDF2
nltk
numpy
scikit-learn
pytest